Divide the g.js monolith into modules.
