+++ src/g.js
@@ -1,0 +1,14 @@
+// g.js
+// JavaScript library for vector graphics
+// https://github.com/nodebox/g.js
+// (c) 2014 EMRG
+// g.js may be freely distributed under the MIT license.
+// Based on: canvas.js, https://github.com/clips/pattern/blob/master/pattern/canvas.js (BSD)
+// De Smedt T. & Daelemans W. (2012). Pattern for Python. Journal of Machine Learning Research.
+
+'use strict';
+
+var g = {};
+g.random = require('random');
+
+module.exports = g;
